requirements.txt
Pillow==10.0.0
numpy==1.24.3
Wand==0.6.11